+++ src/main/java/Utilities/svgIcons.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package Views;
+package Utilities;
 
 import com.formdev.flatlaf.extras.FlatSVGIcon;
 import javax.swing.JLabel;
