+++ src/main/java/Views/Misc/TestingGrounds.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
  */
-package Views;
+package Views.Misc;
 
 import Controllers.Controller;
 import Utilities.Song;
