+++ src/main/java/Views/PlayView.java
@@ -118,7 +118,6 @@ public class PlayView extends javax.swing.JFrame {
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setTitle("Symph [Playing]");
         setBackground(new java.awt.Color(26, 23, 32));
-        setMaximumSize(new java.awt.Dimension(1100, 600));
         setResizable(false);
 
         mainPanel.setBackground(new java.awt.Color(26, 23, 32));
@@ -134,14 +133,10 @@ public class PlayView extends javax.swing.JFrame {
         backButton.setFont(new java.awt.Font("Century Gothic", 1, 36)); // NOI18N
         backButton.setForeground(new java.awt.Color(226, 115, 150));
         backButton.setText("«");
-        backButton.setToolTipText("");
-        backButton.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
 
         forwardButton.setFont(new java.awt.Font("Century Gothic", 1, 36)); // NOI18N
         forwardButton.setForeground(new java.awt.Color(226, 115, 150));
         forwardButton.setText("»");
-        forwardButton.setToolTipText("");
-        forwardButton.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
 
         titleLabel.setFont(new java.awt.Font("Century Gothic", 0, 36)); // NOI18N
         titleLabel.setForeground(new java.awt.Color(240, 240, 240));
@@ -188,8 +183,6 @@ public class PlayView extends javax.swing.JFrame {
         playButton.setFont(new java.awt.Font("Century Gothic", 0, 32)); // NOI18N
         playButton.setForeground(new java.awt.Color(226, 115, 150));
         playButton.setText("►");
-        playButton.setToolTipText("");
-        playButton.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
         playButton.setMaximumSize(new java.awt.Dimension(566, 512));
 
         leaveButton.setFont(new java.awt.Font("Century Gothic", 0, 32)); // NOI18N
@@ -208,8 +201,6 @@ public class PlayView extends javax.swing.JFrame {
         pauseButton.setFont(new java.awt.Font("Century Gothic", 1, 36)); // NOI18N
         pauseButton.setForeground(new java.awt.Color(226, 115, 150));
         pauseButton.setText("II");
-        pauseButton.setToolTipText("");
-        pauseButton.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
 
         javax.swing.GroupLayout playingPanelLayout = new javax.swing.GroupLayout(playingPanel);
         playingPanel.setLayout(playingPanelLayout);
